# api/openapi.yaml
openapi: 3.1.0
info:
  title: sshc Local API
  version: 0.1.0
paths:
  /api/v1/session/bootstrap:
    post:
      operationId: bootstrapSession
      parameters:
        - name: X-SSHC-Bootstrap
          in: header
          required: true
          schema: { type: string, minLength: 43, maxLength: 43 }
      responses:
        "200":
          description: Session established
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BootstrapResponse" }
        "401": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/session/renew:
    post:
      # リロードすると、ページの中にあった CSRF トークンは失われる。そして次の
      # ひとつを生むはずのブートストラップのフラグメントは、初回の使用で消費されて
      # いる。これは、Cookie がすでに名指ししているセッションのためにトークンを
      # 発行する。CSRF トークンを提示しないのは、リロード直後にはそれがないからだ。
      # 守っているのはセッション、Origin、そして Fetch Metadata である。
      operationId: renewSession
      responses:
        "200":
          description: A fresh CSRF token for the existing session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BootstrapResponse" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
...
  /api/v1/remote-keys/plan:
    post:
      operationId: planRemoteKeyRegistration
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RemoteKeyPlanRequest" }
      responses:
        "200":
          description: What the registration would do
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RemoteKeyPlan" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
  /api/v1/remote-keys/register:
    post:
      operationId: registerRemoteKey
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RemoteKeyRegisterRequest" }
      responses:
        "200":
          description: Completed registration
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RemoteKeyRegisterResponse" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
...
  /api/v1/passwords:
    get:
      operationId: getPasswordVault
      responses:
        "200":
          description: Whether a vault exists, whether it is unlocked, and which hosts have a password
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PasswordVaultStatus" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /api/v1/passwords/initialise:
    post:
      operationId: initialisePasswordVault
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PassphraseRequest" }
      responses:
        "200":
          description: Vault created and unlocked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PasswordVaultStatus" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/passwords/unlock:
    post:
      operationId: unlockPasswordVault
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PassphraseRequest" }
      responses:
        "200":
          description: Vault unlocked for this run
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PasswordVaultStatus" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  schemas:
    BootstrapResponse:
      type: object
      additionalProperties: false
      required: [csrfToken]
      properties:
        csrfToken: { type: string, minLength: 43, maxLength: 43 }
...
    Problem:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        detail: { type: string }
        path: { type: string }
        line: { type: integer }
        column: { type: integer }
        diagnostics: { type: array, items: { $ref: "#/components/schemas/Diagnostic" } }
        conflict: { $ref: "#/components/schemas/ConflictReport" }
...
    ExecutableDirective:
      type: object
      additionalProperties: false
      required: [keyword, command, path, line, condition, onEvaluate, onConnect, overridable]
      properties:
        keyword: { type: string }
        command: { type: string }
        path: { type: string }
        line: { type: integer }
        condition: { type: string }
        onEvaluate: { type: boolean }
        onConnect: { type: boolean }
        overridable: { type: boolean }
...
    RemoteKeyPlanRequest:
      type: object
      additionalProperties: false
      required: [alias, keyPath, publicKey]
      properties:
        alias: { type: string, minLength: 1, maxLength: 64 }
        keyPath: { type: string, maxLength: 4096 }
        publicKey: { type: string, minLength: 1, maxLength: 4096 }
    RemoteKeyPlan:
      type: object
      additionalProperties: false
      required: [alias, user, hostname, port, valuesFrom, fingerprint, keyPath, keyLine, remotePath, routine, supported, manual, executableDirectives]
      properties:
        alias: { type: string }
        user: { type: string }
        hostname: { type: string }
        port: { type: string }
        valuesFrom: { type: string }
        fingerprint: { type: string }
        keyPath: { type: string }
        keyLine: { type: string }
        remotePath: { type: string }
        routine: { type: string }
        supported: { type: boolean }
        manual:
          type: array
          items: { type: string }
        executableDirectives:
          type: array
          items: { $ref: "#/components/schemas/ExecutableDirective" }
    RemoteKeyRegisterRequest:
      type: object
      additionalProperties: false
      required: [alias, keyPath, publicKey, acknowledgeExecutable]
      properties:
        alias: { type: string, minLength: 1, maxLength: 64 }
        keyPath: { type: string, maxLength: 4096 }
        publicKey: { type: string, minLength: 1, maxLength: 4096 }
        acknowledgeExecutable: { type: boolean }
    RemoteKeyRegisterResponse:
      type: object
      additionalProperties: false
      required: [outcome, exitCode, stderr, truncated]
      properties:
        outcome: { type: string }
        exitCode: { type: integer }
        stderr: { type: string }
        truncated: { type: boolean }
...
    PasswordVaultStatus:
      type: object
      additionalProperties: false
      required: [exists, unlocked, aliases]
      properties:
        exists: { type: boolean }
        unlocked: { type: boolean }
        # パスワードが保存されているホスト。名前の集合は秘密ではない。秘密なのは
        # 値の方であり、どの操作もそれを返さない。
        aliases:
          type: array
          items: { type: string }
        minPassphraseLength: { type: integer }
    # blocker は保存済みパスワードを使用不能にするので、保存自体を拒否する。
    # warning は、ユーザーの方がよく知っているかもしれない事実である — 設定済みの
    # 鍵が向こう側で認可されていないのはよくあることだ — ので、事実として述べる
    # にとどめ、判断は本来それが属する場所に委ねる。
...
    PassphraseRequest:
      type: object
      additionalProperties: false
      required: [passphrase]
      properties:
        passphrase: { type: string, minLength: 1, maxLength: 1024 }
...
    Diagnostic:
      type: object
      additionalProperties: false
      required: [severity, code]
      properties:
        severity: { type: string }
        code: { type: string }
        path: { type: string }
        absolute: { type: string }
        external: { type: boolean }
        line: { type: integer }
        detail: { type: string }
...
    DiffLine:
      type: object
      additionalProperties: false
      required: [op, text]
      properties:
        op: { type: string }
        text: { type: string }
        oldLine: { type: integer }
        newLine: { type: integer }
...
    ConflictReport:
      type: object
      additionalProperties: false
      required: [path, externalChange, localChange]
      properties:
        path: { type: string }
        baseDigest: { type: string }
        diskDigest: { type: string }
        externalChange: { type: array, items: { $ref: "#/components/schemas/DiffLine" } }
        localChange: { type: array, items: { $ref: "#/components/schemas/DiffLine" } }
...
  responses:
    Problem:
      description: Request rejected
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }